requirements.txt
streamlit
groq
python-dotenv
duckduckgo-search
ddgs